完成支付宝沙盒支付功能

+++ kaiapi-order/src/main/java/com/edkai/order/model/entity/ApiOrder.java
@@ -45,7 +45,7 @@ public class ApiOrder implements Serializable {
     /**
      * 单价
      */
-    private Double charging;
+    private BigDecimal charging;
 
     /**
      * 交易金额
